// src/api/gen/models/AthleteProfile.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AthleteProfile = {
  userId: string;
  goals?: Array<string>;
  level?: string;
  heightCm?: number;
  weightKg?: number;
  createdAt?: string;
  updatedAt?: string;
};